Removed the code necessary for loop and region merging on the C++ side
This should be handled completely within CLIPS because it may turn out that
someone using this library may not need loops and regions to be merged thus the
code has been reverted to what it used to do.

It is now up to CLIPS rules to handle the merging of loops and regions.